+++ src/controllers/statevector_controller.hpp
@@ -15,8 +15,8 @@
 #ifndef _aer_statevector_controller_hpp_
 #define _aer_statevector_controller_hpp_
 
-#include "base/controller.hpp"
-#include "statevector_state.hpp"
+#include "controller.hpp"
+#include "simulators/statevector/statevector_state.hpp"
 
 namespace AER {
 namespace Simulator {
@@ -27,9 +27,9 @@ namespace Simulator {
 
 /**************************************************************************
  * Config settings:
- * 
+ *
  * From Statevector::State class
- * 
+ *
  * - "initial_statevector" (json complex vector): Use a custom initial
  *      statevector for the simulation [Default: null].
  * - "zero_threshold" (double): Threshold for truncating small values to
@@ -42,7 +42,7 @@ namespace Simulator {
  *      measure sampling [Default: 10]
  * - "statevector_hpc_gate_opt" (bool): Enable large qubit gate optimizations.
  *      [Default: False]
- * 
+ *
  * From BaseController Class
  *
  * - "max_parallel_threads" (int): Set the maximum OpenMP threads that may
@@ -55,11 +55,11 @@ namespace Simulator {
  * - "snapshots" (bool): Return snapshots object in circuit data [Default: True]
  * - "memory" (bool): Return memory array in circuit data [Default: False]
  * - "register" (bool): Return register array in circuit data [Default: False]
- * 
+ *
  **************************************************************************/
 
 class StatevectorController : public Base::Controller {
-public:
+ public:
   //-----------------------------------------------------------------------
   // Base class config override
   //-----------------------------------------------------------------------
@@ -70,34 +70,54 @@ public:
   // Allowed config options:
   // - "initial_statevector: complex_vector"
   // Plus Base Controller config options
-  virtual void set_config(const json_t &config) override;
+  virtual void set_config(const json_t& config) override;
 
   // Clear the current config
   void virtual clear_config() override;
 
-protected:
+ protected:
+  virtual size_t required_memory_mb(
+      const Circuit& circuit, const Noise::NoiseModel& noise) const override;
 
-  virtual size_t required_memory_mb(const Circuit& circuit,
-                                    const Noise::NoiseModel& noise) const override;
+  // Simulation methods for the Statevector Controller
+  enum class Method {
+    automatic,
+    statevector_cpu,
+    statevector_thrust_gpu,
+    statevector_thrust_cpu
+  };
 
-private:
+  // Simulation precision
+  enum class Precision { double_precision, single_precision };
 
+ private:
   //-----------------------------------------------------------------------
   // Base class abstract method override
   //-----------------------------------------------------------------------
 
   // This simulator will only return a single shot, regardless of the
   // input shot number
-  virtual ExperimentData run_circuit(const Circuit &circ,
-                                 const Noise::NoiseModel& noise,
-                                 const json_t &config,
-                                 uint_t shots,
-                                 uint_t rng_seed) const override;
+  virtual ExperimentData run_circuit(const Circuit& circ,
+                                     const Noise::NoiseModel& noise,
+                                     const json_t& config, uint_t shots,
+                                     uint_t rng_seed) const override;
 
+  // Execute n-shots of a circuit on the input state
+  template <class State_t>
+  ExperimentData run_circuit_helper(const Circuit& circ,
+                                    const Noise::NoiseModel& noise,
+                                    const json_t& config, uint_t shots,
+                                    uint_t rng_seed) const;
   //-----------------------------------------------------------------------
   // Custom initial state
-  //-----------------------------------------------------------------------        
+  //-----------------------------------------------------------------------
   cvector_t initial_state_;
+
+  // Method for storing statevector
+  Method method_ = Method::automatic;
+
+  // Precision of statevector
+  Precision precision_ = Precision::double_precision;
 };
 
 //=========================================================================
@@ -113,15 +133,41 @@ StatevectorController::StatevectorController() : Base::Controller() {
 // Config
 //-------------------------------------------------------------------------
 
-void StatevectorController::set_config(const json_t &config) {
+void StatevectorController::set_config(const json_t& config) {
   // Set base controller config
   Base::Controller::set_config(config);
 
-  //Add custom initial state
+  // Add custom initial state
   if (JSON::get_value(initial_state_, "initial_statevector", config)) {
     // Check initial state is normalized
     if (!Utils::is_unit_vector(initial_state_, validation_threshold_))
-      throw std::runtime_error("StatevectorController: initial_statevector is not a unit vector");
+      throw std::runtime_error(
+          "StatevectorController: initial_statevector is not a unit vector");
+  }
+
+  // Add method
+  std::string method;
+  if (JSON::get_value(method, "method", config)) {
+    if (method == "statevector" || method == "statevector_cpu") {
+      method_ = Method::statevector_cpu;
+    } else if (method == "statevector_gpu") {
+      method_ = Method::statevector_thrust_gpu;
+    } else if (method == "statevector_thrust") {
+      method_ = Method::statevector_thrust_cpu;
+    } else if (method != "automatic") {
+      throw std::runtime_error(
+          std::string("UnitaryController: Invalid simulation method (") +
+          method + std::string(")."));
+    }
+  }
+
+  std::string precision;
+  if (JSON::get_value(precision, "precision", config)) {
+    if (precision == "double") {
+      precision_ = Precision::double_precision;
+    } else if (precision == "single") {
+      precision_ = Precision::single_precision;
+    }
   }
 }
 
@@ -130,21 +176,87 @@ void StatevectorController::clear_config() {
   initial_state_ = cvector_t();
 }
 
-size_t StatevectorController::required_memory_mb(const Circuit& circ,
-                                                 const Noise::NoiseModel& noise) const {
-  Statevector::State<> state;
-  return state.required_memory_mb(circ.num_qubits, circ.ops);
+size_t StatevectorController::required_memory_mb(
+    const Circuit& circ, const Noise::NoiseModel& noise) const {
+  if (precision_ == Precision::single_precision) {
+    Statevector::State<QV::QubitVector<float>> state;
+    return state.required_memory_mb(circ.num_qubits, circ.ops);
+  } else {
+    Statevector::State<> state;
+    return state.required_memory_mb(circ.num_qubits, circ.ops);
+  }
 }
 
 //-------------------------------------------------------------------------
 // Run circuit
 //-------------------------------------------------------------------------
 
-ExperimentData StatevectorController::run_circuit(const Circuit &circ,
-                                              const Noise::NoiseModel& noise,
-                                              const json_t &config,
-                                              uint_t shots,
-                                              uint_t rng_seed) const {
+ExperimentData StatevectorController::run_circuit(
+    const Circuit& circ, const Noise::NoiseModel& noise, const json_t& config,
+    uint_t shots, uint_t rng_seed) const {
+  switch (method_) {
+    case Method::automatic:
+    case Method::statevector_cpu: {
+      if (precision_ == Precision::double_precision) {
+        // Double-precision Statevector simulation
+        return run_circuit_helper<Statevector::State<QV::QubitVector<double>>>(
+            circ, noise, config, shots, rng_seed);
+      } else {
+        // Single-precision Statevector simulation
+        return run_circuit_helper<Statevector::State<QV::QubitVector<float>>>(
+            circ, noise, config, shots, rng_seed);
+      }
+    }
+    case Method::statevector_thrust_gpu: {
+#ifdef AER_THRUST_CUDA
+      if (precision_ == Precision::double_precision) {
+        // Double-precision Statevector simulation
+        return run_circuit_helper<
+            Statevector::State<QV::QubitVectorThrust<double>>>(
+            circ, noise, config, shots, rng_seed);
+      } else {
+        // Single-precision Statevector simulation
+        return run_circuit_helper<
+            Statevector::State<QV::QubitVectorThrust<float>>>(
+            circ, noise, config, shots, rng_seed);
+      }
+#else
+      throw std::runtime_error(
+          "StatevectorController: method statevector_gpu is not supported on "
+          "this "
+          "system");
+#endif
+    }
+    case Method::statevector_thrust_cpu: {
+#ifdef AER_THRUST_CPU
+      if (precision_ == Precision::double_precision) {
+        // Double-precision Statevector simulation
+        return run_circuit_helper<
+            Statevector::State<QV::QubitVectorThrust<double>>>(
+            circ, noise, config, shots, rng_seed);
+      } else {
+        // Single-precision Statevector simulation
+        return run_circuit_helper<
+            Statevector::State<QV::QubitVectorThrust<float>>>(
+            circ, noise, config, shots, rng_seed);
+      }
+#else
+      throw std::runtime_error(
+          "StatevectorController: method statevector_thrust is not supported "
+          "on this "
+          "system");
+#endif
+    }
+    default:
+      throw std::runtime_error(
+          "StatevectorController:Invalid simulation method");
+  }
+}
+
+template <class State_t>
+ExperimentData StatevectorController::run_circuit_helper(
+    const Circuit& circ, const Noise::NoiseModel& noise, const json_t& config,
+    uint_t shots, uint_t rng_seed) const {
   // Initialize  state
   Statevector::State<> state;
 
@@ -165,7 +277,7 @@ ExperimentData StatevectorController::run_circuit(const Circuit &circ,
   // Set config
   state.set_config(config);
   state.set_parallalization(parallel_state_update_);
-  
+
   // Rng engine
   RngEngine rng;
   rng.set_seed(rng_seed);
@@ -173,7 +285,7 @@ ExperimentData StatevectorController::run_circuit(const Circuit &circ,
   // Output data container
   ExperimentData data;
   data.set_config(config);
-  
+
   // Run single shot collecting measure data or snapshots
   if (initial_state_.empty())
     state.initialize_qreg(circ.num_qubits);
@@ -182,7 +294,7 @@ ExperimentData StatevectorController::run_circuit(const Circuit &circ,
   state.initialize_creg(circ.num_memory, circ.num_registers);
   state.apply_ops(circ.ops, data, rng);
   state.add_creg_to_data(data);
-  
+
   // Add final state to the data
   data.add_additional_data("statevector", state.qreg().vector());
 
@@ -190,8 +302,8 @@ ExperimentData StatevectorController::run_circuit(const Circuit &circ,
 }
 
 //-------------------------------------------------------------------------
-} // end namespace Simulator
+}  // end namespace Simulator
 //-------------------------------------------------------------------------
-} // end namespace AER
+}  // end namespace AER
 //-------------------------------------------------------------------------
 #endif