+++ test/src/test_snapshot_bdd.cpp
@@ -2,7 +2,7 @@
 #include <map>
 #include <catch.hpp>
 
-#include <simulators/qasm/qasm_controller.hpp>
+#include <controllers/qasm_controller.hpp>
 
 #include "utils.hpp"